requirements.txt
requests==2.25.1
Flask==1.1.2
python-rapidjson==1.0
